+++ Server/R/HIVModelManager.R
@@ -95,18 +95,27 @@ HIVModelManager <- R6::R6Class(
         PrintAlert('Starting HIV Model main fit task')
 
         private$Catalogs$MainFitTask <- Task$new(
-          function(caseData, aggrData, settings, parameters, popCombination, aggrDataSelection) {
+          function(
+            caseData,
+            aggrData,
+            settings,
+            parameters,
+            popCombination,
+            aggrDataSelection
+          ) {
             suppressMessages(pkgload::load_all())
             options(width = 120)
 
+            PrintH1('Parameters')
+
             if (length(popCombination$Aggr) > 0) {
-              PrintH1('Aggregated data selection')
+              PrintH2('Aggregated data selection')
               cli::cli_text(
                 '{.val {cat(capture.output(aggrDataSelection), sep = "\n")}}'
               )
             }
 
-            PrintH1('Population combination')
+            PrintH2('Population combination')
             if (!is.null(caseData)) {
               if (length(popCombination$CaseAbbr) > 0) {
                 PrintAlert('Case-based populations: {.val {popCombination$CaseAbbr}}')
@@ -123,12 +132,12 @@ HIVModelManager <- R6::R6Class(
               PrintAlert('Aggregated populations: {.val \'None\'}')
             }
 
-            PrintH1('Time intervals and diagnosis rates modelling')
+            PrintH2('Time intervals and diagnosis rates modelling')
             cli::cli_text(
               '{.val {cat(capture.output(parameters$Intervals), sep = "\n")}}'
             )
 
-            PrintH1('Advanced paramaters')
+            PrintH2('Advanced paramaters')
             PrintAlert(' 1. Range of calculations                                : {.val {parameters$ModelMinYear} - {parameters$ModelMaxYear}}')
             PrintAlert(' 2. HIV diagnoses, total                                 : {.val {parameters$FitPosMinYear} - {parameters$FitPosMaxYear}}')
             PrintAlert(' 3. HIV diagnoses, by CD4 count                          : {.val {parameters$FitPosCD4MinYear} - {parameters$FitPosCD4MaxYear}}')
@@ -142,18 +151,23 @@ HIVModelManager <- R6::R6Class(
             PrintAlert('11. Extra diagnosis rate due to non-AIDS symptoms        : {.val {parameters$Delta4Fac}}')
             PrintAlert('12. Country-specific settings                            : {.val {parameters$Country}}')
 
-            dataSets <- hivEstimatesAccuracy2::CombineData(
-              caseData, aggrData, popCombination, aggrDataSelection
-            )
+            dataSets <- CombineData(caseData, aggrData, popCombination, aggrDataSelection)
+
+            PrintH1('Performing main fit')
 
             impResult <- list()
             for (imp in names(dataSets)) {
+              PrintH2('Iteration {.val {imp}}')
               context <- hivModelling::GetRunContext(
                 data = dataSets[[imp]],
                 settings = settings,
-                parameters = parameters
+                parameters = list(
+                  INCIDENCE = parameters
+                )
               )
               popData <- hivModelling::GetPopulationData(context)
+
+              print(popData)
 
               startTime <- Sys.time()
               fitResults <- hivModelling::PerformMainFit(
@@ -173,7 +187,7 @@ HIVModelManager <- R6::R6Class(
               )
             }
 
-            plotData <- hivEstimatesAccuracy2::GetHIVPlotData(impResult, NULL)
+            plotData <- GetHIVPlotData(impResult, NULL)
 
             result <- list(
               MainFitResult = impResult,
@@ -191,15 +205,6 @@ HIVModelManager <- R6::R6Class(
             aggrDataSelection = aggrDataSelection
           ),
           session = private$Session,
-          progressCallback = function(runLog) {
-            private$SendMessage(
-              'MODELS_RUN_LOG_SET',
-              payload = list(
-                ActionStatus = 'SUCCESS',
-                RunLog = runLog
-              )
-            )
-          },
           successCallback = function(result) {
             private$Catalogs$PopCombination <- popCombination
             private$Catalogs$AggrDataSelection <- aggrDataSelection
@@ -216,7 +221,7 @@ HIVModelManager <- R6::R6Class(
               )
             )
           },
-          failCallback = function(msg = null) {
+          failCallback = function(msg = NULL) {
             if (!is.null(msg)) {
               PrintAlert(msg, type = 'danger')
             }
@@ -304,9 +309,11 @@ HIVModelManager <- R6::R6Class(
           ) {
             suppressMessages(pkgload::load_all())
             options(width = 120)
+
             mainCount <- length(mainFitResult)
             fits <- list()
             i <- 0
+
             for (imp in names(mainFitResult)) {
               i <- i + 1
               mainFit <- mainFitResult[[imp]]
@@ -415,19 +422,21 @@ HIVModelManager <- R6::R6Class(
             bsCount = bsCount,
             bsType = bsType,
             maxRunTime = maxRunTime,
-            mainFitResult = private$Catalogs$MainFitResult,
-            caseData = private$AppMgr$CaseMgr$Data,
-            aggrData = private$AppMgr$AggrMgr$Data,
-            popCombination = private$Catalogs$PopCombination,
-            aggrDataSelection = private$Catalogs$AggrDataSelection
+            mainFitResult = isolate(private$Catalogs$MainFitResult),
+            caseData = isolate(private$AppMgr$CaseMgr$Data),
+            aggrData = isolate(private$AppMgr$AggrMgr$Data),
+            popCombination = isolate(private$Catalogs$PopCombination),
+            aggrDataSelection = isolate(private$Catalogs$AggrDataSelection)
           ),
           session = private$Session,
           successCallback = function(result) {
+            PrintAlert('Running HIV Model bootstrap fit task finished')
+
             private$Catalogs$BootstrapFitResult <- result$Fits
             private$Catalogs$BootstrapFitStats <- result$Stats
             private$Catalogs$PlotData <- result$PlotData
             private$InvalidateAfterStep('BOOTSTRAP')
-            PrintAlert('Running HIV Model bootstrap fit task finished')
+
             private$SendMessage(
               'BOOTSTRAP_RUN_FINISHED',
               payload = list(
@@ -438,10 +447,11 @@ HIVModelManager <- R6::R6Class(
             )
           },
           failCallback = function(msg = NULL) {
+            PrintAlert('Running HIV Model bootstrap fit task failed', type = 'danger')
             if (!is.null(msg)) {
               PrintAlert(msg, type = 'danger')
             }
-            PrintAlert('Running HIV Model bootstrap fit task failed', type = 'danger')
+            private$Catalogs$BootstrapFitTask$Stop(force = TRUE)
             private$SendMessage(
               'BOOTSTRAP_RUN_FINISHED',
               payload = list(
@@ -451,6 +461,7 @@ HIVModelManager <- R6::R6Class(
             )
           }
         )
+
         private$SendMessage(
           'BOOTSTRAP_RUN_STARTED',
           payload = list(
@@ -475,6 +486,7 @@ HIVModelManager <- R6::R6Class(
 
     CancelBootstrapFit = function() {
       if (!is.null(private$Catalogs$BootstrapFitTask)) {
+        private$Catalogs$BootstrapFitTask$Stop()
         private$SendMessage(
           'BOOTSTRAP_RUN_CANCELLED',
           payload = list(