+++ Server/R/Messages.R
@@ -74,6 +74,7 @@ PrintAlert <- function(
     'warning' = cli::cli_alert_warning,
     'info'    = cli::cli_alert_info,
     'success' = cli::cli_alert_success,
+    'text'    = cli::cli_text,
     cli::cli_alert
   )